decoupled env.ts, redirect.ts, transport.ts files. Moved into a common folder that will be used as common code between CDK V1 and V2
