+++ src/common/redirect.ts
@@ -6,13 +6,16 @@
  * Copyright 2021 Datadog, Inc.
  */
 
-import * as lambda from "@aws-cdk/aws-lambda";
 import log from "loglevel";
-import { RuntimeType, runtimeLookup } from "./index";
-export const DD_HANDLER_ENV_VAR = "DD_LAMBDA_HANDLER";
-export const PYTHON_HANDLER = "datadog_lambda.handler.handler";
-export const JS_HANDLER_WITH_LAYERS = "/opt/nodejs/node_modules/datadog-lambda-js/handler.handler";
-export const JS_HANDLER = "node_modules/datadog-lambda-js/dist/handler.handler";
+import {
+  LambdaFunction,
+  RuntimeType,
+  runtimeLookup,
+  DD_HANDLER_ENV_VAR,
+  JS_HANDLER_WITH_LAYERS,
+  JS_HANDLER,
+  PYTHON_HANDLER,
+} from "../index";
 
 /**
  * To avoid modifying code in the user's lambda handler, redirect the handler to a Datadog
@@ -21,10 +24,11 @@ export const JS_HANDLER = "node_modules/datadog-lambda-js/dist/handler.handler";
  * replacement handler to find.
  */
 
-export function redirectHandlers(lambdas: lambda.Function[], addLayers: boolean) {
+export function redirectHandlers(lambdas: LambdaFunction[], addLayers: boolean) {
   log.debug(`Wrapping Lambda function handlers with Datadog handler...`);
   lambdas.forEach((lam) => {
-    const cfnFunction = lam.node.defaultChild as lambda.CfnFunction;
+    // const cfnFunction = lam.node.defaultChild as lambda.CfnFunction;
+    const cfnFunction = lam.node.defaultChild;
     const originalHandler = cfnFunction.handler as string;
     lam.addEnvironment(DD_HANDLER_ENV_VAR, originalHandler);
     const handler = getDDHandler(lam, addLayers);
@@ -36,7 +40,7 @@ export function redirectHandlers(lambdas: lambda.Function[], addLayers: boolean)
   });
 }
 
-function getDDHandler(lam: lambda.Function, addLayers: boolean) {
+function getDDHandler(lam: LambdaFunction, addLayers: boolean) {
   const runtime: string = lam.runtime.name;
   const lambdaRuntime: RuntimeType = runtimeLookup[runtime];
   if (lambdaRuntime === undefined || lambdaRuntime === RuntimeType.UNSUPPORTED) {
