+++ src/common/transport.ts
@@ -6,9 +6,9 @@
  * Copyright 2021 Datadog, Inc.
  */
 
-import * as lambda from "@aws-cdk/aws-lambda";
 import log from "loglevel";
-import { runtimeLookup, RuntimeType } from ".";
+import { runtimeLookup, RuntimeType } from "..";
+import { LambdaFunction } from "../index";
 
 export const API_KEY_ENV_VAR = "DD_API_KEY";
 export const API_KEY_SECRET_ARN_ENV_VAR = "DD_API_KEY_SECRET_ARN";
@@ -64,7 +64,7 @@ export class Transport {
     this.apiKmsKey = apiKmsKey;
   }
 
-  applyEnvVars(lambdas: lambda.Function[]) {
+  applyEnvVars(lambdas: LambdaFunction[]) {
     log.debug(`Setting Datadog transport environment variables...`);
     lambdas.forEach((lam) => {
       lam.addEnvironment(FLUSH_METRICS_TO_LOGS_ENV_VAR, this.flushMetricsToLogs.toString());
