+++ src/datadog.ts
@@ -12,31 +12,19 @@ import * as lambdaPython from "@aws-cdk/aws-lambda-python";
 import * as logs from "@aws-cdk/aws-logs";
 import * as cdk from "@aws-cdk/core";
 import log from "loglevel";
-import { applyLayers, redirectHandlers, addForwarder, addForwarderToLogGroups, applyEnvVariables } from "./index";
-import { Transport } from "./transport";
+import { Transport } from "./common/transport";
+import {
+  applyLayers,
+  redirectHandlers,
+  addForwarder,
+  addForwarderToLogGroups,
+  applyEnvVariables,
+  DatadogProps,
+  validateProps,
+  TagKeys,
+} from "./index";
+
 const versionJson = require("../version.json");
-
-export interface DatadogProps {
-  readonly pythonLayerVersion?: number;
-  readonly nodeLayerVersion?: number;
-  readonly extensionLayerVersion?: number;
-  readonly addLayers?: boolean;
-  readonly forwarderArn?: string;
-  readonly flushMetricsToLogs?: boolean;
-  readonly site?: string;
-  readonly apiKey?: string;
-  readonly apiKeySecretArn?: string;
-  readonly apiKmsKey?: string;
-  readonly enableDatadogTracing?: boolean;
-  readonly injectLogContext?: boolean;
-  readonly logLevel?: string;
-  readonly enableDatadogLogs?: boolean;
-  readonly captureLambdaPayload?: boolean;
-}
-
-enum TagKeys {
-  Cdk = "dd_cdk_construct",
-}
 
 export const defaultProps = {
   addLayers: true,
@@ -155,55 +143,4 @@ export function addCdkConstructVersionTag(lambdaFunctions: lambda.Function[]) {
       includeResourceTypes: ["AWS::Lambda::Function"],
     });
   });
-}
-
-function validateProps(props: DatadogProps) {
-  log.debug("Validating props...");
-
-  checkForMultipleApiKeys(props);
-  const siteList: string[] = [
-    "datadoghq.com",
-    "datadoghq.eu",
-    "us3.datadoghq.com",
-    "us5.datadoghq.com",
-    "ddog-gov.com",
-  ];
-  if (props.site !== undefined && !siteList.includes(props.site.toLowerCase())) {
-    throw new Error(
-      "Warning: Invalid site URL. Must be either datadoghq.com, datadoghq.eu, us3.datadoghq.com, us5.datadoghq.com, or ddog-gov.com.",
-    );
-  }
-
-  if (
-    props.apiKey === undefined &&
-    props.apiKmsKey === undefined &&
-    props.apiKeySecretArn === undefined &&
-    props.flushMetricsToLogs === false
-  ) {
-    throw new Error(
-      "When `flushMetricsToLogs` is false, `apiKey`, `apiKeySecretArn`, or `apiKmsKey` must also be set.",
-    );
-  }
-  if (props.extensionLayerVersion !== undefined) {
-    if (props.apiKey === undefined && props.apiKeySecretArn === undefined && props.apiKmsKey === undefined) {
-      throw new Error("When `extensionLayer` is set, `apiKey`, `apiKeySecretArn`, or `apiKmsKey` must also be set.");
-    }
-  }
-}
-
-export function checkForMultipleApiKeys(props: DatadogProps) {
-  let multipleApiKeysMessage;
-  if (props.apiKey !== undefined && props.apiKmsKey !== undefined && props.apiKeySecretArn !== undefined) {
-    multipleApiKeysMessage = "`apiKey`, `apiKmsKey`, and `apiKeySecretArn`";
-  } else if (props.apiKey !== undefined && props.apiKmsKey !== undefined) {
-    multipleApiKeysMessage = "`apiKey` and `apiKmsKey`";
-  } else if (props.apiKey !== undefined && props.apiKeySecretArn !== undefined) {
-    multipleApiKeysMessage = "`apiKey` and `apiKeySecretArn`";
-  } else if (props.apiKmsKey !== undefined && props.apiKeySecretArn !== undefined) {
-    multipleApiKeysMessage = "`apiKmsKey` and `apiKeySecretArn`";
-  }
-
-  if (multipleApiKeysMessage) {
-    throw new Error(`${multipleApiKeysMessage} should not be set at the same time.`);
-  }
 }
