+++ src/layer.ts
@@ -6,42 +6,12 @@
  * Copyright 2021 Datadog, Inc.
  */
 
-import * as crypto from "crypto";
 import * as lambda from "@aws-cdk/aws-lambda";
 import { Architecture } from "@aws-cdk/aws-lambda";
 import * as cdk from "@aws-cdk/core";
 import log from "loglevel";
-export const DD_ACCOUNT_ID = "464622532012";
-export const DD_GOV_ACCOUNT_ID = "002406178527";
-
-export enum RuntimeType {
-  NODE,
-  PYTHON,
-  UNSUPPORTED,
-}
-const layerPrefix = "DatadogLayer";
-const extensionLayerPrefix = "DatadogExtension";
-export const runtimeLookup: { [key: string]: RuntimeType } = {
-  "nodejs10.x": RuntimeType.NODE,
-  "nodejs12.x": RuntimeType.NODE,
-  "nodejs14.x": RuntimeType.NODE,
-  "python2.7": RuntimeType.PYTHON,
-  "python3.6": RuntimeType.PYTHON,
-  "python3.7": RuntimeType.PYTHON,
-  "python3.8": RuntimeType.PYTHON,
-  "python3.9": RuntimeType.PYTHON,
-};
-
-const runtimeToLayerName: { [key: string]: string } = {
-  "nodejs10.x": "Datadog-Node10-x",
-  "nodejs12.x": "Datadog-Node12-x",
-  "nodejs14.x": "Datadog-Node14-x",
-  "python2.7": "Datadog-Python27",
-  "python3.6": "Datadog-Python36",
-  "python3.7": "Datadog-Python37",
-  "python3.8": "Datadog-Python38",
-  "python3.9": "Datadog-Python39",
-};
+import * as constants from "./common/constants";
+import * as exp from "./common/experiment";
 
 const layers: Map<string, lambda.ILayerVersion> = new Map();
 
@@ -58,38 +28,38 @@ export function applyLayers(
   log.debug("Applying layers to Lambda functions...");
   lambdas.forEach((lam) => {
     const runtime: string = lam.runtime.name;
-    const lambdaRuntimeType: RuntimeType = runtimeLookup[runtime];
+    const lambdaRuntimeType: constants.RuntimeType = constants.runtimeLookup[runtime];
     const isARM = lam.architecture === Architecture.ARM_64;
-    const isNode = lambdaRuntimeType === RuntimeType.NODE;
-    if (lambdaRuntimeType === RuntimeType.UNSUPPORTED) {
+    const isNode = lambdaRuntimeType === constants.RuntimeType.NODE;
+    if (lambdaRuntimeType === constants.RuntimeType.UNSUPPORTED) {
       log.debug(`Unsupported runtime: ${runtime}`);
       return;
     }
 
     let lambdaLayerArn;
     let extensionLayerArn;
-    if (lambdaRuntimeType === RuntimeType.PYTHON) {
+    if (lambdaRuntimeType === constants.RuntimeType.PYTHON) {
       if (pythonLayerVersion === undefined) {
-        errors.push(getMissingLayerVersionErrorMsg(lam.node.id, "Python", "python"));
+        errors.push(exp.getMissingLayerVersionErrorMsg(lam.node.id, "Python", "python"));
         return;
       }
-      lambdaLayerArn = getLambdaLayerArn(region, pythonLayerVersion, runtime, isARM, isNode);
+      lambdaLayerArn = exp.getLambdaLayerArn(region, pythonLayerVersion, runtime, isARM, isNode);
       log.debug(`Using Python Lambda layer: ${lambdaLayerArn}`);
       addLayer(lambdaLayerArn, false, scope, lam, runtime);
     }
 
-    if (lambdaRuntimeType === RuntimeType.NODE) {
+    if (lambdaRuntimeType === constants.RuntimeType.NODE) {
       if (nodeLayerVersion === undefined) {
-        errors.push(getMissingLayerVersionErrorMsg(lam.node.id, "Node.js", "node"));
+        errors.push(exp.getMissingLayerVersionErrorMsg(lam.node.id, "Node.js", "node"));
         return;
       }
-      lambdaLayerArn = getLambdaLayerArn(region, nodeLayerVersion, runtime, isARM, isNode);
+      lambdaLayerArn = exp.getLambdaLayerArn(region, nodeLayerVersion, runtime, isARM, isNode);
       log.debug(`Using Node Lambda layer: ${lambdaLayerArn}`);
       addLayer(lambdaLayerArn, false, scope, lam, runtime);
     }
 
     if (extensionLayerVersion !== undefined) {
-      extensionLayerArn = getExtensionLayerArn(region, extensionLayerVersion, isARM);
+      extensionLayerArn = exp.getExtensionLayerArn(region, extensionLayerVersion, isARM);
       log.debug(`Using extension layer: ${extensionLayerArn}`);
       addLayer(extensionLayerArn, true, scope, lam, runtime);
     }
@@ -106,9 +76,9 @@ function addLayer(
 ) {
   let layerId;
   if (isExtensionLayer) {
-    layerId = generateExtensionLayerId(lam.functionArn);
+    layerId = exp.generateExtensionLayerId(lam.functionArn);
   } else {
-    layerId = generateLambdaLayerId(lam.functionArn, runtime);
+    layerId = exp.generateLambdaLayerId(lam.functionArn, runtime);
   }
 
   if (layerArn !== undefined) {
@@ -120,48 +90,4 @@ function addLayer(
     log.debug(`Adding layer ${lambdaLayer} to Lambda: ${lam.functionName}`);
     lam.addLayers(lambdaLayer);
   }
-}
-
-export function getLambdaLayerArn(region: string, version: number, runtime: string, isArm: boolean, isNode: boolean) {
-  const baseLayerName = runtimeToLayerName[runtime];
-  const layerName = isArm && !isNode ? `${baseLayerName}-ARM` : baseLayerName;
-  // TODO: edge case where gov cloud is the region, but they are using a token so we can't resolve it.
-  const isGovCloud = region === "us-gov-east-1" || region === "us-gov-west-1";
-
-  // if this is a GovCloud region, use the GovCloud lambda layer
-  if (isGovCloud) {
-    log.debug("GovCloud region detected, using the GovCloud lambda layer");
-    return `arn:aws-us-gov:lambda:${region}:${DD_GOV_ACCOUNT_ID}:layer:${layerName}:${version}`;
-  }
-  return `arn:aws:lambda:${region}:${DD_ACCOUNT_ID}:layer:${layerName}:${version}`;
-}
-
-export function getExtensionLayerArn(region: string, version: number, isArm: boolean) {
-  const baseLayerName = "Datadog-Extension";
-  const layerName = isArm ? `${baseLayerName}-ARM` : baseLayerName;
-  const isGovCloud = region === "us-gov-east-1" || region === "us-gov-west-1";
-  if (isGovCloud) {
-    log.debug("GovCloud region detected, using the GovCloud extension layer");
-    return `arn:aws-us-gov:lambda:${region}:${DD_GOV_ACCOUNT_ID}:layer:${layerName}:${version}`;
-  }
-  return `arn:aws:lambda:${region}:${DD_ACCOUNT_ID}:layer:${layerName}:${version}`;
-}
-
-export function getMissingLayerVersionErrorMsg(functionKey: string, formalRuntime: string, paramRuntime: string) {
-  return (
-    `Resource ${functionKey} has a ${formalRuntime} runtime, but no ${formalRuntime} Lambda Library version was provided. ` +
-    `Please add the '${paramRuntime}LayerVersion' parameter for the Datadog serverless macro.`
-  );
-}
-
-export function generateLambdaLayerId(lambdaFunctionArn: string, runtime: string) {
-  log.debug("Generating construct Id for Datadog Lambda layer");
-  const layerValue: string = crypto.createHash("sha256").update(lambdaFunctionArn).digest("hex");
-  return layerPrefix + "-" + runtime + "-" + layerValue;
-}
-
-export function generateExtensionLayerId(lambdaFunctionArn: string) {
-  log.debug("Generating construct Id for Datadog Extension layer");
-  const layerValue: string = crypto.createHash("sha256").update(lambdaFunctionArn).digest("hex");
-  return extensionLayerPrefix + "-" + layerValue;
 }
