+++ test/index.spec.ts
@@ -3,14 +3,16 @@ import * as cdk from "@aws-cdk/core";
 import "@aws-cdk/assert/jest";
 import {
   Datadog,
+  DD_ACCOUNT_ID,
   FLUSH_METRICS_TO_LOGS_ENV_VAR,
   ENABLE_DD_TRACING_ENV_VAR,
   INJECT_LOG_CONTEXT_ENV_VAR,
   ENABLE_DD_LOGS_ENV_VAR,
   CAPTURE_LAMBDA_PAYLOAD_ENV_VAR,
+  JS_HANDLER_WITH_LAYERS,
+  DD_HANDLER_ENV_VAR,
+  PYTHON_HANDLER,
 } from "../src/index";
-import { DD_ACCOUNT_ID } from "../src/layer";
-import { JS_HANDLER_WITH_LAYERS, DD_HANDLER_ENV_VAR, PYTHON_HANDLER } from "../src/redirect";
 import { findDatadogSubscriptionFilters } from "./test-utils";
 
 describe("addLambdaFunctions", () => {
