+++ test/transport.spec.ts
@@ -12,8 +12,8 @@ import {
   ENABLE_DD_LOGS_ENV_VAR,
   API_KEY_SECRET_ARN_ENV_VAR,
   CAPTURE_LAMBDA_PAYLOAD_ENV_VAR,
+  DD_HANDLER_ENV_VAR,
 } from "../src/index";
-import { DD_HANDLER_ENV_VAR } from "../src/redirect";
 const EXTENSION_LAYER_VERSION = 5;
 
 describe("SITE_URL_ENV_VAR", () => {
